requirements.txt
fastapi[standard]~=0.115.10
pydantic~=2.10.6
pydantic-settings~=2.8.1

pymongo~=4.11.1
beanie~=1.29.0

bcrypt~=4.3.0
pyjwt~=2.10.1
